refactor: split image convert files
